mask and move significance methods to bias class
